Refactor: break down pipelineb test cases to smaller files
